requirements.txt
torch>=2.0.0
torchtext>=0.15.0
datasets>=2.10.0
huggingface-hub>=0.15.0
matplotlib>=3.5.0
tqdm>=4.64.0
tensorboard>=2.10.0
pyyaml>=6.0
numpy>=1.21.0
sentencepiece>=0.1.97  # 用于BPE分词